Renamed original client.cpp to tcp_client.cpp. Keeping it for back compatibility with ROS1. Not tested or being used.
